% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-direction.R
\name{.ambiguous_control_types}
\alias{.ambiguous_control_types}
\title{control_type values che sono intrinsecamente ambigui sul role assignment}
\usage{
.ambiguous_control_types
}
\description{
control_type values che sono intrinsecamente ambigui sul role assignment
}
\keyword{internal}
