% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-build.R
\name{.annotate_direction}
\alias{.annotate_direction}
\title{Annota records pair con direction_check}
\usage{
.annotate_direction(records_pair)
}
\description{
Annota records pair con direction_check
}
\keyword{internal}
