% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-contrast-anchor.R
\name{.ca_candidates}
\alias{.ca_candidates}
\title{Candidati da passare al resolver, dal piu' specifico al piu' generico}
\usage{
.ca_candidates(
  values,
  treated_label = NA_character_,
  contrast_class = NA_character_
)
}
\arguments{
\item{values}{valori TRATTATI delle chiavi che cambiano (la classe dominante)}

\item{treated_label}{etichetta leggibile del braccio trattato}

\item{contrast_class}{classe dominante del delta}
}
\description{
Candidati da passare al resolver, dal piu' specifico al piu' generico
}
\keyword{internal}
