% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-contrast-anchor.R
\name{.ca_drop_doses}
\alias{.ca_drop_doses}
\title{Toglie dosi e tempi PRESERVANDO i separatori}
\usage{
.ca_drop_doses(x)
}
\description{
Serve al rilevatore di combinazioni, che deve ancora vedere \code{/} e
\code{_}. \code{.ca_strip_units()} invece trasforma \code{/} in spazio (li'
e' voluto: \code{ug/ml} e' rumore), e usarla qui faceva sparire il separatore
prima dello split: \code{Bleomycin/Alpha-Lipoic Acid} diventava la sola
bleomicina (regressione misurata dall'equivalenza 2026-07-27).
}
\keyword{internal}
