% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.canon_json}
\alias{.canon_json}
\title{Serializzazione JSON canonica ordine-insensibile}
\usage{
.canon_json(x)
}
\description{
Oggetti: chiavi ordinate. Array: elementi serializzati e POI ordinati (tutti
gli array della firma sono insiemi, nessun ordine significativo). Scalari:
gia' normalizzati a stringa, quotati.
}
\keyword{internal}
