% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-contrast-gate.R
\name{.CG_DRUGCLASS_RX}
\alias{.CG_DRUGCLASS_RX}
\title{Nome di CLASSE farmacologica: un bersaglio piu' il ruolo, non una molecola}
\usage{
.CG_DRUGCLASS_RX
}
\description{
"MEK inhibitor" identifica una famiglia, non un composto: quando il resolver
l'ha agganciato a U0126 ha messo insieme inibitori di MEK, ERK e JNK
(censimento 2026-07-28).
}
\keyword{internal}
