% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-contrast-gate.R
\name{.cg_normalize}
\alias{.cg_normalize}
\title{Normalizza i separatori: "_" e "|" non sono lettere}
\usage{
.cg_normalize(x)
}
\description{
Normalizza i separatori: "_" e "|" non sono lettere
}
\keyword{internal}
