% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-coherence.R
\name{.cluster_coherence_signals}
\alias{.cluster_coherence_signals}
\title{Riassume i segnali di coerenza (Fase A/C) da un data.frame di contrasti}
\usage{
.cluster_coherence_signals(contrast_df)
}
\description{
Riassume i segnali di coerenza (Fase A/C) da un data.frame di contrasti
}
\keyword{internal}
