% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-counts-cache.R
\name{.default_stage4_cache_dir}
\alias{.default_stage4_cache_dir}
\title{Default cache dir per Stadio 4 counts}
\usage{
.default_stage4_cache_dir()
}
\description{
Default cache dir per Stadio 4 counts
}
\keyword{internal}
