% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgx-utils.R
\name{.dgx_run_id_short}
\alias{.dgx_run_id_short}
\title{Estrae lo slug user-defined da un run_id pieno}
\usage{
.dgx_run_id_short(run_id)
}
\arguments{
\item{run_id}{stringa formato \verb{<ts>-<slug>-<hex>}.}
}
\value{
slug come character(1).
}
\description{
Estrae lo slug user-defined da un run_id pieno
}
\keyword{internal}
