% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-name-recovery.R
\name{.extract_agent_term}
\alias{.extract_agent_term}
\title{Estrae il termine-agente (composto) dai metadati}
\usage{
.extract_agent_term(source, characteristics, title)
}
\description{
Estrae il termine-agente (composto) dai metadati
}
\keyword{internal}
