% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-counts-cache.R
\name{.h5_gene_axis}
\alias{.h5_gene_axis}
\title{Gene axis (HGNC symbol, make.unique) di un H5, memoizzato}
\usage{
.h5_gene_axis(h5_path)
}
\description{
Gene axis (HGNC symbol, make.unique) di un H5, memoizzato
}
\keyword{internal}
