% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-counts-cache.R
\name{.h5_sample_axis}
\alias{.h5_sample_axis}
\title{Sample axis (geo_accession) di un H5, memoizzato}
\usage{
.h5_sample_axis(h5_path)
}
\description{
Sample axis (geo_accession) di un H5, memoizzato
}
\keyword{internal}
