% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-summary.R
\name{.isTRUE_vec}
\alias{.isTRUE_vec}
\title{Versione vettorizzata di isTRUE che gestisce NA in modo sicuro}
\usage{
.isTRUE_vec(x)
}
\arguments{
\item{x}{vettore logico (puo' contenere NA)}
}
\value{
vettore logical(length(x)) con TRUE solo dove x e' esattamente TRUE
}
\description{
Versione vettorizzata di isTRUE che gestisce NA in modo sicuro
}
\keyword{internal}
