% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.json_quote}
\alias{.json_quote}
\title{Quoting JSON di una stringa (escape minimale, deterministico)}
\usage{
.json_quote(s)
}
\description{
Quoting JSON di una stringa (escape minimale, deterministico)
}
\keyword{internal}
