% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-technical-lanes.R
\name{.lane_index}
\alias{.lane_index}
\title{Estrae l'indice di corsia dal titolo}
\usage{
.lane_index(x)
}
\arguments{
\item{x}{character, i titoli GEO.}
}
\value{
integer, l'indice di corsia o NA se il titolo non ne dichiara uno.
}
\description{
Estrae l'indice di corsia dal titolo
}
\keyword{internal}
