% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-technical-lanes.R
\name{.lane_strip}
\alias{.lane_strip}
\title{Toglie dal titolo i marcatori di corsia}
\usage{
.lane_strip(x)
}
\arguments{
\item{x}{character, i titoli GEO.}
}
\value{
character, il titolo senza il marcatore di corsia.
}
\description{
Toglie dal titolo i marcatori di corsia
}
\keyword{internal}
