% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-dispatch.R
\name{.lookup_rg}
\alias{.lookup_rg}
\title{Lookup replicate_group da study via group_id}
\usage{
.lookup_rg(study, group_id)
}
\description{
Lookup replicate_group da study via group_id
}
\keyword{internal}
