% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-shard-merge.R
\name{.merge_attr}
\alias{.merge_attr}
\title{Unisce i data.frame di un attributo su piu' pezzi}
\usage{
.merge_attr(shards, tab, nome)
}
\description{
Unisce i data.frame di un attributo su piu' pezzi
}
\keyword{internal}
