% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.norm_dose}
\alias{.norm_dose}
\title{Token dose: value_numeric+unit se presente, altrimenti fallback value_raw}
\usage{
.norm_dose(dose)
}
\description{
Token dose: value_numeric+unit se presente, altrimenti fallback value_raw
}
\keyword{internal}
