% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.norm_engmod}
\alias{.norm_engmod}
\title{Normalizza una modifica ingegnerizzata (esclude la description libera)}
\usage{
.norm_engmod(em)
}
\description{
Normalizza una modifica ingegnerizzata (esclude la description libera)
}
\keyword{internal}
