% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.norm_scalar}
\alias{.norm_scalar}
\title{Normalizza uno scalare a stringa canonica NA-aware}
\usage{
.norm_scalar(x)
}
\description{
NULL, NA, length-0 e stringhe vuote (dopo trim) collassano sul token
assente \code{""}. Tutto il resto diventa il valore trimmato come carattere.
}
\keyword{internal}
