% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage2-design-signature.R
\name{.norm_set}
\alias{.norm_set}
\title{Normalizza un set di stringhe (array) -> lista di token, vuota se assente}
\usage{
.norm_set(x)
}
\description{
Normalizza un set di stringhe (array) -> lista di token, vuota se assente
}
\keyword{internal}
