% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-row-pairing.R
\name{.rp_normalize_separators}
\alias{.rp_normalize_separators}
\title{Normalizza i separatori prima di applicare le espressioni regolari}
\usage{
.rp_normalize_separators(x)
}
\description{
\code{_} e \code{|} sono separatori, non lettere: per le espressioni regolari
\code{_} e' pero' carattere di parola, quindi \code{\\bsirna\\b} non vede
\code{Control_siRNA_1} e \code{10day} dentro \code{osimertinib_10day} non e'
un tempo. Stesso inciampo gia' pagato su \code{calcium_low}.
}
\keyword{internal}
