% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-row-pairing.R
\name{.rp_time_mismatch}
\alias{.rp_time_mismatch}
\title{I due bracci dichiarano tempi diversi e nessuno in comune}
\usage{
.rp_time_mismatch(treated_label, control_label)
}
\description{
I due bracci dichiarano tempi diversi e nessuno in comune
}
\keyword{internal}
