% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-shard-merge.R
\name{.SHARD_ATTR_PSD}
\alias{.SHARD_ATTR_PSD}
\title{Registri che viaggiano come attributi delle due tabelle}
\usage{
.SHARD_ATTR_PSD
}
\description{
Registri che viaggiano come attributi delle due tabelle
}
\keyword{internal}
