% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-name-recovery.R
\name{.slugify}
\alias{.slugify}
\title{Produce uno slug ASCII minuscolo (spazi e punteggiatura -> underscore)}
\usage{
.slugify(x)
}
\description{
Produce uno slug ASCII minuscolo (spazi e punteggiatura -> underscore)
}
\keyword{internal}
