% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-dispatch.R
\name{.split_record_id}
\alias{.split_record_id}
\title{Split record_id in (series_id, suffix) sul primo "__"}
\usage{
.split_record_id(record_id)
}
\description{
I record_id sono \code{sprintf("\%s__\%s", series_id, comparison_id|group_id)}
(\code{R/stage3-build.R}). I comparison_id / group_id possono contenere
ulteriori \code{"__"}: l'unica garanzia e' che \code{series_id} non contiene
\code{"__"} (GSE seguono il pattern \code{GSE<n>}).
}
\keyword{internal}
