% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage3-io.R
\name{load_stage3}
\alias{load_stage3}
\title{Legge stage3_result da directory scritta con \code{write_stage3_to_dir()}}
\usage{
load_stage3(dir)
}
\arguments{
\item{dir}{path directory scritta da \code{\link{write_stage3_to_dir}}.}
}
\value{
\code{stage3_result} S3 object con le 5 componenti.
}
\description{
Ricostruisce il \code{stage3_result} S3 leggendo i 5 file nella directory:
\code{assignments.parquet}, \code{clusters.rds}, \code{record_summary.rds},
\code{non_clusterable.rds}, \code{run_metadata.json}.
}
\seealso{
\code{\link{write_stage3_to_dir}}
}
