% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-io.R
\name{load_stage4}
\alias{load_stage4}
\title{Carica stage4 result da directory}
\usage{
load_stage4(dir)
}
\arguments{
\item{dir}{directory contenente i 5 file Stage 4.}
}
\value{
list stage4_result.
}
\description{
Ricostruisce un oggetto \code{stage4_result} dai 5 file scritti da
\code{write_stage4_to_dir}. Round-trip preserva per_study_de,
cluster_pooled, qc_report, non_processable, config e run_metadata
(run_id + timestamp).
}
