% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-stats.R
\name{mcnemar_paired}
\alias{mcnemar_paired}
\title{McNemar test paired su due classificatori}
\usage{
mcnemar_paired(pred_a, pred_b, continuity = FALSE)
}
\arguments{
\item{pred_a}{character vector con valori "correct"/"wrong" del primo classificatore}

\item{pred_b}{character vector con valori "correct"/"wrong" del secondo classificatore}

\item{continuity}{applica correzione di continuita' (Yates), default FALSE}
}
\value{
list con statistic, p_value, b, c, n
}
\description{
Confronta due classificatori sulle stesse osservazioni rispetto a un
gold standard implicito (ognuno e' "correct" o "wrong"). Calcola b
(A correct, B wrong) e c (A wrong, B correct), poi statistica chi^2 =
(b-c)^2 / (b+c). Sample con NA in entrambi i predittori esclusi.
}
\details{
Equivalente a mcnemar.test(table(pred_a, pred_b), correct = continuity).
}
