% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{sha256_text}
\alias{sha256_text}
\title{Calcola lo SHA-256 di una stringa testuale UTF-8}
\usage{
sha256_text(x)
}
\arguments{
\item{x}{stringa di lunghezza 1}
}
\value{
stringa esadecimale di 64 caratteri
}
\description{
Calcola lo SHA-256 di una stringa testuale UTF-8
}
\keyword{internal}
