% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage4-dispatch.R
\name{.lookup_cmp}
\alias{.lookup_cmp}
\title{Lookup comparison da study via comparison_id}
\usage{
.lookup_cmp(study, comparison_id)
}
\description{
Lookup comparison da study via comparison_id
}
\keyword{internal}
