% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulomicsr-package.R
\docType{package}
\name{simulomicsr-package}
\alias{simulomicsr}
\alias{simulomicsr-package}
\title{simulomicsr: Easy simulator of OMICS data in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulate OMICS data in a simple way (i.e., without any parameter specification), for a quick models prototyping and tests. PROS: works out-of-the-box without setting (quite) any additional parameter. CONS: not suitable for complex designs.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Luca Vedovelli \email{luca.vedovelli@ubep.unipd.it} (\href{https://orcid.org/0000-0003-4847-2333}{ORCID})
}

}
\keyword{internal}
