% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulomicsr-package.R
\docType{package}
\name{simulomicsr-package}
\alias{simulomicsr}
\alias{simulomicsr-package}
\title{simulomicsr: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Luca Vedovelli \email{luca.vedovelli@ubep.unipd.it} (\href{https://orcid.org/0000-0003-4847-2333}{ORCID})
}

}
\keyword{internal}
